+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 - **fixed:** panic message on formatting failure to mention that, instead of a wrong "capacity overflow"
+- **fixed:** `(try_)alloc_try_with` UB when allocating inside the provided closure, memory leak when using an *unallocated* bump allocator
+- **breaking:** `(try_)alloc_try_with` now requires a guaranteed allocated `Bump(Scope)`
+- **added:** `(try_)alloc_try_with_mut` as an optimized version of `(try_)alloc_try_with`
+- **improved:** `reset_to` now takes a `&` instead of `&mut`
 
 ## 0.7.0 (2024-08-31)
 - **breaking:** allow returning values from closure arguments of `scoped`, `aligned` and `scoped_aligned`